+++ crates/ruma-common/src/push/iter.rs
@@ -53,9 +53,16 @@ impl AnyPushRule {
         self.as_ref().triggers_highlight()
     }
 
-    /// Whether an event that matches the push rule should trigger a notification.
+    /// Whether an event that matches the push rule should trigger a notification (either in-app or
+    /// remote / push).
     pub fn triggers_notification(&self) -> bool {
         self.as_ref().triggers_notification()
+    }
+
+    /// Whether an event that matches the push rule should trigger a remote notification.
+    #[cfg(feature = "unstable-msc3768")]
+    pub fn triggers_remote_notification(&self) -> bool {
+        self.as_ref().triggers_remote_notification()
     }
 
     /// The sound that should be played when an event matches the push rule, if any.
@@ -182,9 +189,16 @@ impl<'a> AnyPushRuleRef<'a> {
         self.actions().iter().any(|a| a.is_highlight())
     }
 
-    /// Whether an event that matches the push rule should trigger a notification.
+    /// Whether an event that matches the push rule should trigger a notification (either in-app or
+    /// remote / push).
     pub fn triggers_notification(self) -> bool {
         self.actions().iter().any(|a| a.should_notify())
+    }
+
+    /// Whether an event that matches the push rule should trigger a remote notification.
+    #[cfg(feature = "unstable-msc3768")]
+    pub fn triggers_remote_notification(self) -> bool {
+        self.actions().iter().any(|a| a.should_notify_remote())
     }
 
     /// The sound that should be played when an event matches the push rule, if any.